Cliniqo/requirements.txt
streamlit>=1.35.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
bcrypt>=4.1.0
gspread>=6.0.0
google-auth>=2.29.0
